+++ setup.py
@@ -37,7 +37,6 @@ setup(
             "rich",
             "pytest",
             "pytest-cov",
-            "scikit-build",
         ]
     }
 )
