+++ src/picologging/__init__.py
@@ -408,62 +408,3 @@ class FileHandler(StreamHandler):
     def __repr__(self):
         level = getLevelName(self.level)
         return '<%s %s (%s)>' % (self.__class__.__name__, self.baseFilename, level)
-
-
-class WatchedFileHandler(FileHandler):
-    """
-    A handler for logging to a file, which watches the file
-    to see if it has changed while in use. This can happen because of
-    usage of programs such as newsyslog and logrotate which perform
-    log file rotation. This handler, intended for use under Unix,
-    watches the file to see if it has changed since the last emit.
-    (A file has changed if its device or inode have changed.)
-    If it has changed, the old file stream is closed, and the file
-    opened to get a new stream.
-
-    This handler is not appropriate for use under Windows, because
-    under Windows open files cannot be moved or renamed - logging
-    opens the files with exclusive locks - and so there is no need
-    for such a handler. Furthermore, ST_INO is not supported under
-    Windows; stat always returns zero for this value.
-    """
-    def __init__(self, filename, mode='a', encoding=None, delay=False):
-        super().__init__(filename, mode=mode, encoding=encoding, delay=delay)
-        self.dev, self.ino = -1, -1
-        self._statstream()
-
-    def _statstream(self):
-        if self.stream:
-            sres = os.fstat(self.stream.fileno())
-            self.dev, self.ino = sres.st_dev, sres.st_ino
-
-    def reopenIfNeeded(self):
-        """
-        Reopen log file if needed.
-
-        Checks if the underlying file has changed, and if it
-        has, close the old stream and reopen the file to get the
-        current stream.
-        """
-        try:
-            sres = os.stat(self.baseFilename)
-        except FileNotFoundError:
-            sres = None
-
-        if not sres or sres.st_dev != self.dev or sres.st_ino != self.ino:
-            if self.stream is not None:
-                self.stream.flush()
-                self.stream.close()
-                self.stream = None  # See Issue #21742: _open () might fail.
-                self.stream = self._open()
-                self._statstream()
-
-    def emit(self, record):
-        """
-        Emit a record.
-
-        If underlying file has changed, reopen the file before emitting the
-        record to it.
-        """
-        self.reopenIfNeeded()
-        FileHandler.emit(self, record)
